Make each user_inputs struct self-contained
each parameter set has a declare_parameters() and an assign_parameters() method acting on/reading from a dealii::ParameterHandler.
