+++ include/prismspf/user_inputs/constraint_parameters.h
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include <deal.II/base/parameter_handler.h>
 #include <deal.II/base/point.h>
 #include <deal.II/base/tensor.h>
 #include <deal.II/base/types.h>
@@ -184,6 +185,20 @@ public:
     return false;
   }
 
+  /**
+   * @brief Declare the parameters to be read from an input file.
+   */
+  void
+  declare_parameters(dealii::ParameterHandler &parameter_handler,
+                     unsigned int              max_criteria = 5) const;
+
+  /**
+   * @brief Assign the parameters read from an input file to this object.
+   */
+  void
+  assign_parameters(dealii::ParameterHandler &parameter_handler,
+                    unsigned int              max_criteria = 5);
+
   // Map of boundary conditions. The first key is the field index.
   std::unordered_map<std::string, FieldConstraints<dim>> boundary_condition_list;
 };
@@ -210,6 +225,151 @@ BoundaryParameters<dim>::print_parameter_summary() const
   // TODO print pinned points
 
   ConditionalOStreams::pout_summary() << "\n" << std::flush;
+}
+
+template <unsigned int dim>
+inline void
+BoundaryParameters<dim>::declare_parameters(dealii::ParameterHandler &parameter_handler,
+                                            unsigned int              max_criteria) const
+{
+  for (unsigned int criterion_id = 0; criterion_id < max_criteria; criterion_id++)
+    {
+      std::string subsection_text =
+        "boundary conditions: " + std::to_string(criterion_id);
+      parameter_handler.enter_subsection(subsection_text);
+      {
+        parameter_handler.declare_entry(
+          "variables",
+          "",
+          dealii::Patterns::Anything(),
+          "The names of the fields that will use these constraints.");
+        parameter_handler.declare_entry("conditions",
+                                        "",
+                                        dealii::Patterns::Anything(),
+                                        "List of conditions.");
+        parameter_handler.enter_subsection("pinning point");
+        {
+          parameter_handler.declare_entry("enable pinned point",
+                                          "false",
+                                          dealii::Patterns::Bool(),
+                                          "Whether to use a pinned point");
+
+          parameter_handler.declare_entry("x",
+                                          "0.0",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "X-coordinate of the point");
+          parameter_handler.declare_entry("y",
+                                          "0.0",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Y-coordinate of the point");
+          parameter_handler.declare_entry("z",
+                                          "0.0",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Z-coordinate of the point");
+
+          parameter_handler.declare_entry("value",
+                                          "2147483647",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Value of pinned point.");
+
+          parameter_handler.declare_entry("x value",
+                                          "2147483647",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Value of pinned point for the x-component.");
+          parameter_handler.declare_entry("y value",
+                                          "2147483647",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Value of pinned point for the y-component.");
+          parameter_handler.declare_entry("z value",
+                                          "2147483647",
+                                          dealii::Patterns::Double(-DBL_MAX, DBL_MAX),
+                                          "Value of pinned point for the z-component.");
+        }
+        parameter_handler.leave_subsection();
+      }
+      parameter_handler.leave_subsection();
+    }
+}
+
+template <unsigned int dim>
+inline void
+BoundaryParameters<dim>::assign_parameters(dealii::ParameterHandler &parameter_handler,
+                                           unsigned int              max_criteria)
+{
+  static const std::vector<std::string> axis_labels = {"x", "y", "z"};
+  for (unsigned int criterion_id = 0; criterion_id < max_criteria; criterion_id++)
+    {
+      std::string subsection_text =
+        "boundary conditions: " + std::to_string(criterion_id);
+      parameter_handler.enter_subsection(subsection_text);
+      {
+        std::vector<std::string> field_names =
+          dealii::Utilities::split_string_list(parameter_handler.get("variables"));
+        std::vector<std::string> conditions_strings =
+          dealii::Utilities::split_string_list(parameter_handler.get("conditions"));
+
+        ComponentConditions component_conditions;
+        if (conditions_strings.size() == 1)
+          {
+            // all the same
+            for (unsigned int boundary_id = 0; boundary_id < 2 * dim; boundary_id++)
+              {
+                component_conditions.conditions[boundary_id] =
+                  condition_from_string(conditions_strings[0]);
+              }
+          }
+        else
+          {
+            for (unsigned int boundary_id = 0; boundary_id < conditions_strings.size();
+                 boundary_id++)
+              {
+                component_conditions.conditions[boundary_id] =
+                  condition_from_string(conditions_strings[boundary_id]);
+              }
+          }
+
+        // Attach conditions to fields
+        for (const auto &field_comp_name : field_names)
+          {
+            int                    pos = field_comp_name.length() - 2;
+            const std::string      end = field_comp_name.substr(pos > 0 ? pos : 0);
+            std::string            field_name;
+            std::set<unsigned int> comps;
+            if (end == ":x")
+              {
+                comps      = {0};
+                field_name = field_comp_name.substr(0, pos);
+              }
+            else if (end == ":y")
+              {
+                comps      = {1};
+                field_name = field_comp_name.substr(0, pos);
+              }
+            else if (end == ":z")
+              {
+                comps      = {2};
+                field_name = field_comp_name.substr(0, pos);
+              }
+            else
+              {
+                for (unsigned int comp = 0; comp < dim; ++comp)
+                  {
+                    comps.insert(comp);
+                  }
+                field_name = field_comp_name;
+              }
+            for (unsigned int component : comps)
+              {
+                if (component < dim)
+                  {
+                    boundary_condition_list[field_name].component_constraints.at(
+                      component) = component_conditions;
+                  }
+              }
+          }
+      }
+      parameter_handler.leave_subsection();
+    }
 }
 
 PRISMS_PF_END_NAMESPACE
