+++ src/user_inputs/user_input_parameters.cc
@@ -1,571 +1,9 @@
 // SPDX-FileCopyrightText: © 2025 PRISMS Center at the University of Michigan
 // SPDX-License-Identifier: GNU Lesser General Public Version 2.1
 
-#include <deal.II/base/config.h>
-#include <deal.II/base/parameter_handler.h>
-#include <deal.II/base/point.h>
-#include <deal.II/base/utilities.h>
-
-#include <boost/algorithm/string/predicate.hpp>
-
-#include <prismspf/core/exceptions.h>
-#include <prismspf/core/grid_refiner_criterion.h>
-#include <prismspf/core/type_enums.h>
-#include <prismspf/core/types.h>
-
-#include <prismspf/user_inputs/constraint_parameters.h>
-#include <prismspf/user_inputs/input_file_reader.h>
-#include <prismspf/user_inputs/linear_solve_parameters.h>
-#include <prismspf/user_inputs/load_initial_condition_parameters.h>
-#include <prismspf/user_inputs/nonlinear_solve_parameters.h>
-#include <prismspf/user_inputs/temporal_discretization.h>
 #include <prismspf/user_inputs/user_input_parameters.h>
 
-#include <prismspf/config.h>
-
-#include <climits>
-#include <cmath>
-#include <string>
-#include <vector>
-
 PRISMS_PF_BEGIN_NAMESPACE
-
-template <unsigned int dim>
-UserInputParameters<dim>::UserInputParameters(const std::string &file_name)
-{
-  InputFileReader           input_file_reader(file_name);
-  dealii::ParameterHandler &parameter_handler = input_file_reader.get_parameter_handler();
-  // Assign the parameters to the appropriate data structures
-  assign_spatial_discretization_parameters(parameter_handler);
-  assign_temporal_discretization_parameters(parameter_handler);
-  assign_linear_solve_parameters(parameter_handler);
-  assign_nonlinear_solve_parameters(parameter_handler);
-  assign_output_parameters(parameter_handler);
-  assign_checkpoint_parameters(parameter_handler);
-  assign_boundary_parameters(parameter_handler);
-  assign_load_initial_condition_parameters(parameter_handler);
-  assign_nucleation_parameters(parameter_handler);
-  assign_miscellaneous_parameters(parameter_handler);
-  load_model_constants(input_file_reader, parameter_handler);
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_spatial_discretization_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  // Rectangular mesh
-  parameter_handler.enter_subsection("Rectangular mesh");
-  {
-    static const std::vector<std::string> axis_labels = {"x", "y", "z"};
-    RectangularMesh<dim>                 &rect = spatial_discretization.rectangular_mesh;
-    for (unsigned int i = 0; i < dim; ++i)
-      {
-        rect.size[i] = parameter_handler.get_double(axis_labels[i] + " size");
-        rect.lower_bound[i] =
-          parameter_handler.get_double(axis_labels[i] + " lower bound");
-        rect.subdivisions[i] = static_cast<unsigned int>(
-          parameter_handler.get_integer(axis_labels[i] + " subdivisions"));
-        if (parameter_handler.get_bool("periodic " + axis_labels[i]))
-          {
-            rect.periodic_directions.insert(i);
-          }
-      }
-  }
-  parameter_handler.leave_subsection();
-  // Spherical mesh
-  parameter_handler.enter_subsection("Spherical mesh");
-  {
-    spatial_discretization.spherical_mesh.radius = parameter_handler.get_double("radius");
-  }
-  parameter_handler.leave_subsection();
-
-  spatial_discretization.global_refinement =
-    (static_cast<unsigned int>(parameter_handler.get_integer("global refinement")));
-
-  spatial_discretization.has_adaptivity = (parameter_handler.get_bool("mesh adaptivity"));
-
-  spatial_discretization.remeshing_period =
-    (static_cast<unsigned int>(parameter_handler.get_integer("remeshing period")));
-
-  spatial_discretization.max_refinement =
-    (static_cast<unsigned int>(parameter_handler.get_integer("max refinement")));
-  spatial_discretization.min_refinement =
-    (static_cast<unsigned int>(parameter_handler.get_integer("min refinement")));
-
-  for (unsigned int criterion_id = 0; criterion_id < InputFileReader::max_criteria;
-       criterion_id++)
-    {
-      std::string subsection_text =
-        "refinement criterion: " + std::to_string(criterion_id);
-      parameter_handler.enter_subsection(subsection_text);
-      {
-        std::vector<std::string> field_names =
-          dealii::Utilities::split_string_list(parameter_handler.get("variables"));
-        static const std::map<std::string, RefinementFlags> crit_map {
-          {"",                   RefinementFlags::Nothing                          },
-          {"none",               RefinementFlags::Nothing                          },
-          {"value",              RefinementFlags::Value                            },
-          {"gradient",           RefinementFlags::Gradient                         },
-          {"value_and_gradient", RefinementFlags::Value | RefinementFlags::Gradient}
-        };
-        // todo: make case insensitive
-        RefinementCriterion criterion(crit_map.at(parameter_handler.get("type")),
-                                      parameter_handler.get_double("value lower bound"),
-                                      parameter_handler.get_double("value upper bound"),
-                                      parameter_handler.get_double(
-                                        "gradient magnitude lower bound"));
-        for (const auto &field_name : field_names)
-          {
-            spatial_discretization.refinement_criteria[field_name] = criterion;
-          }
-      }
-      parameter_handler.leave_subsection();
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_temporal_discretization_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  temporal_discretization.dt = parameter_handler.get_double("time step");
-  temporal_discretization.num_increments =
-    static_cast<unsigned int>(parameter_handler.get_integer("final increment"));
-  double final_time = parameter_handler.get_double("end time");
-  if (final_time > 0.0)
-    {
-      temporal_discretization.num_increments =
-        std::ceil(final_time / temporal_discretization.dt);
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_output_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  parameter_handler.enter_subsection("output");
-  {
-    output_parameters.directory = parameter_handler.get("directory");
-    output_parameters.file_name = parameter_handler.get("file name");
-    output_parameters.file_type = parameter_handler.get("file type");
-    output_parameters.patch_subdivisions =
-      static_cast<unsigned int>(parameter_handler.get_integer("subdivisions"));
-
-    std::string  condition = parameter_handler.get("condition");
-    unsigned int num_outputs =
-      static_cast<unsigned int>(parameter_handler.get_integer("number"));
-    unsigned int num_increments = temporal_discretization.num_increments;
-    if (condition == "EQUAL_SPACING")
-      {
-        output_parameters.add_equal_spacing_outputs(num_outputs, num_increments);
-      }
-    else if (condition == "LOG_SPACING")
-      {
-        output_parameters.add_log_spacing_outputs(num_outputs, num_increments);
-      }
-    else if (condition == "N_PER_DECADE")
-      {
-        output_parameters.add_n_per_decade_outputs(num_outputs, num_increments);
-      }
-    output_parameters.add_output_list(dealii::Utilities::string_to_int(
-      dealii::Utilities::split_string_list(parameter_handler.get("list"))));
-
-    output_parameters.print_output_period =
-      static_cast<unsigned int>(parameter_handler.get_integer("print step period"));
-    output_parameters.print_timing_with_output =
-      parameter_handler.get_bool("timing information with output");
-  }
-  parameter_handler.leave_subsection();
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_checkpoint_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  parameter_handler.enter_subsection("checkpoints");
-  {
-    checkpoint_parameters.should_load_checkpoint =
-      parameter_handler.get_bool("load from checkpoint");
-
-    std::string  condition = parameter_handler.get("condition");
-    unsigned int num_checkpoints =
-      static_cast<unsigned int>(parameter_handler.get_integer("number"));
-    unsigned int num_increments = temporal_discretization.num_increments;
-    if (condition == "EQUAL_SPACING")
-      {
-        checkpoint_parameters.add_equal_spacing_checkpoints(num_checkpoints,
-                                                            num_increments);
-      }
-    else if (condition == "LOG_SPACING")
-      {
-        checkpoint_parameters.add_log_spacing_checkpoints(num_checkpoints,
-                                                          num_increments);
-      }
-    else if (condition == "N_PER_DECADE")
-      {
-        checkpoint_parameters.add_n_per_decade_checkpoints(num_checkpoints,
-                                                           num_increments);
-      }
-    checkpoint_parameters.add_checkpoint_list(dealii::Utilities::string_to_int(
-      dealii::Utilities::split_string_list(parameter_handler.get("list"))));
-  }
-  parameter_handler.leave_subsection();
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_boundary_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  static const std::vector<std::string> axis_labels = {"x", "y", "z"};
-  for (unsigned int criterion_id = 0; criterion_id < InputFileReader::max_criteria;
-       criterion_id++)
-    {
-      std::string subsection_text =
-        "boundary conditions: " + std::to_string(criterion_id);
-      parameter_handler.enter_subsection(subsection_text);
-      {
-        std::vector<std::string> field_names =
-          dealii::Utilities::split_string_list(parameter_handler.get("variables"));
-        std::vector<std::string> conditions_strings =
-          dealii::Utilities::split_string_list(parameter_handler.get("conditions"));
-
-        ComponentConditions component_conditions;
-        if (conditions_strings.size() == 1)
-          {
-            // all the same
-            for (unsigned int boundary_id = 0; boundary_id < 2 * dim; boundary_id++)
-              {
-                component_conditions.conditions[boundary_id] =
-                  condition_from_string(conditions_strings[0]);
-              }
-          }
-        else
-          {
-            for (unsigned int boundary_id = 0; boundary_id < conditions_strings.size();
-                 boundary_id++)
-              {
-                component_conditions.conditions[boundary_id] =
-                  condition_from_string(conditions_strings[boundary_id]);
-              }
-          }
-
-        // Attach conditions to fields
-        for (const auto &field_comp_name : field_names)
-          {
-            int                    pos = field_comp_name.length() - 2;
-            const std::string      end = field_comp_name.substr(pos > 0 ? pos : 0);
-            std::string            field_name;
-            std::set<unsigned int> comps;
-            if (end == ":x")
-              {
-                comps      = {0};
-                field_name = field_comp_name.substr(0, pos);
-              }
-            else if (end == ":y")
-              {
-                comps      = {1};
-                field_name = field_comp_name.substr(0, pos);
-              }
-            else if (end == ":z")
-              {
-                comps      = {2};
-                field_name = field_comp_name.substr(0, pos);
-              }
-            else
-              {
-                for (unsigned int comp = 0; comp < dim; ++comp)
-                  {
-                    comps.insert(comp);
-                  }
-                field_name = field_comp_name;
-              }
-            for (unsigned int component : comps)
-              {
-                if (component < dim)
-                  {
-                    boundary_parameters.boundary_condition_list[field_name]
-                      .component_constraints.at(component) = component_conditions;
-                  }
-              }
-          }
-      }
-      parameter_handler.leave_subsection();
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_linear_solve_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  for (unsigned int criterion_id = 0; criterion_id < InputFileReader::max_criteria;
-       criterion_id++)
-    {
-      // For linear solves
-      std::string subsection_text =
-        "linear solver parameters: " + std::to_string(criterion_id);
-      parameter_handler.enter_subsection(subsection_text);
-      {
-        std::vector<int> solver_ids = dealii::Utilities::string_to_int(
-          dealii::Utilities::split_string_list(parameter_handler.get("solver_ids")));
-
-        LinearSolverParameters linear_solver_parameters;
-
-        // Set the linear solver type
-        linear_solver_parameters.solver_type = parameter_handler.get("solver type");
-
-        // Set the tolerance type
-        static const std::map<std::string, SolverToleranceType> tolerance_types = {
-          {"AbsoluteResidual",   AbsoluteResidual  },
-          {"RMSEPerField",       RMSEPerField      },
-          {"IntegratedPerField", IntegratedPerField},
-          {"RMSETotal",          RMSETotal         },
-          {"IntegratedTotal",    IntegratedTotal   },
-        };
-        const std::string type_string = parameter_handler.get("tolerance type");
-        const auto       &iter        = tolerance_types.find(type_string);
-        if (iter != tolerance_types.end())
-          {
-            linear_solver_parameters.tolerance_type = iter->second;
-          }
-        else
-          {
-            AssertThrow(false, UnreachableCode());
-          }
-
-        // Set the tolerance value
-        linear_solver_parameters.tolerance =
-          parameter_handler.get_double("tolerance value");
-
-        // Set the maximum number of iterations
-        linear_solver_parameters.max_iterations =
-          static_cast<unsigned int>(parameter_handler.get_integer("max iterations"));
-
-        // Set preconditioner type and related parameters
-        static const std::map<std::string, PreconditionerType> preconditioner_map = {
-          {"None",      None     },
-          {"none",      None     },
-          {"",          None     },
-          {"Chebyshev", Chebyshev},
-          {"chebyshev", Chebyshev},
-          {"GMG",       GMG      },
-          {"gmg",       GMG      }
-        };
-        linear_solver_parameters.preconditioner =
-          preconditioner_map.at(parameter_handler.get("preconditioner type"));
-
-        parameter_handler.enter_subsection("Chebyshev");
-        {
-          linear_solver_parameters.chebyshev_parameters.degree =
-            static_cast<unsigned int>(parameter_handler.get_integer("smoother degree"));
-
-          linear_solver_parameters.chebyshev_parameters.smoothing_range =
-            parameter_handler.get_double("smoothing range");
-
-          linear_solver_parameters.chebyshev_parameters.eig_cg_n_iterations =
-            static_cast<unsigned int>(
-              parameter_handler.get_integer("eigenvalue cg iterations"));
-        }
-        parameter_handler.leave_subsection();
-        parameter_handler.enter_subsection("Richardson");
-        {
-          linear_solver_parameters.richardson_parameters.omega =
-            parameter_handler.get_double("omega");
-          linear_solver_parameters.richardson_parameters.use_preconditioned_residual =
-            parameter_handler.get_bool("use preconditioned residual");
-        }
-        parameter_handler.leave_subsection();
-        parameter_handler.enter_subsection("BiCGStab");
-        {
-          linear_solver_parameters.bicgstab_parameters.exact_residual =
-            parameter_handler.get_bool("exact residual");
-          linear_solver_parameters.bicgstab_parameters.breakdown =
-            parameter_handler.get_double("breakdown");
-        }
-        parameter_handler.leave_subsection();
-        parameter_handler.enter_subsection("GMRES");
-        {
-          using OStrat = dealii::LinearAlgebra::OrthogonalizationStrategy;
-          static const std::map<std::string, OStrat> orthogonalization_strategy_names = {
-            {"classical_gram_schmidt",         OStrat::classical_gram_schmidt        },
-            {"modified_gram_schmidt",          OStrat::modified_gram_schmidt         },
-            {"delayed_classical_gram_schmidt", OStrat::delayed_classical_gram_schmidt}
-          };
-          linear_solver_parameters.gmres_parameters.orthogonalization_strategy =
-            orthogonalization_strategy_names.at(
-              parameter_handler.get("orthogonalization strategy"));
-          linear_solver_parameters.gmres_parameters.right_preconditioning =
-            parameter_handler.get_bool("right preconditioning");
-          linear_solver_parameters.gmres_parameters.use_default_residual =
-            parameter_handler.get_bool("use default residual");
-          linear_solver_parameters.gmres_parameters.force_re_orthogonalization =
-            parameter_handler.get_bool("force re-orthogonalization");
-          linear_solver_parameters.gmres_parameters.batched_mode =
-            parameter_handler.get_bool("batched mode");
-        }
-        parameter_handler.leave_subsection();
-
-        linear_solver_parameters.min_mg_level =
-          static_cast<unsigned int>(parameter_handler.get_integer("min mg level"));
-        for (auto solver_id : solver_ids)
-          {
-            linear_solve_parameters.linear_solvers[static_cast<unsigned int>(solver_id)] =
-              linear_solver_parameters;
-          }
-      }
-      parameter_handler.leave_subsection();
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_nonlinear_solve_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  for (unsigned int criterion_id = 0; criterion_id < InputFileReader::max_criteria;
-       criterion_id++)
-    {
-      // For newton solves
-      std::string subsection_text =
-        "newton solver parameters: " + std::to_string(criterion_id);
-      parameter_handler.enter_subsection(subsection_text);
-      {
-        std::vector<int> solver_ids = dealii::Utilities::string_to_int(
-          dealii::Utilities::split_string_list(parameter_handler.get("solver_ids")));
-        NonlinearSolverParameters nonlinear_solver_parameters;
-        nonlinear_solver_parameters.max_iterations =
-          static_cast<unsigned int>(parameter_handler.get_integer("max iterations"));
-        nonlinear_solver_parameters.step_length =
-          parameter_handler.get_double("step size");
-        nonlinear_solver_parameters.tolerance_value =
-          parameter_handler.get_double("tolerance value");
-        for (auto solver_id : solver_ids)
-          {
-            nonlinear_solve_parameters
-              .newton_solvers[static_cast<unsigned int>(solver_id)] =
-              nonlinear_solver_parameters;
-          }
-      }
-      parameter_handler.leave_subsection();
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_load_initial_condition_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  load_ic_parameters.set_read_initial_conditions_from_file(
-    parameter_handler.get_bool("read initial conditions from file"));
-  static std::array<std::string, 3> axis_labels = {
-    {"x", "y", "z"}
-  };
-  for (unsigned int i = 0; i < Numbers::max_subsections; i++)
-    {
-      parameter_handler.enter_subsection("initial condition file " + std::to_string(i));
-      {
-        // Check if the file is specified
-        if (!parameter_handler.get("file name").empty())
-          {
-            // Create the LoadICFile object
-            InitialConditionFile ic_file;
-            ic_file.filename              = parameter_handler.get("file name");
-            const std::string type_string = parameter_handler.get("dataset format");
-            bool              found_type  = false;
-            for (unsigned int j = 0;
-                 j < static_cast<unsigned int>(DataFormatType::LastEntry);
-                 j++)
-              {
-                if (boost::iequals(type_string,
-                                   to_string(static_cast<DataFormatType>(j))))
-                  {
-                    ic_file.dataset_format = static_cast<DataFormatType>(j);
-                    found_type             = true;
-                    break;
-                  }
-              }
-            AssertThrow(found_type,
-                        dealii::ExcMessage("Unsupported dataset format: " + type_string));
-            ic_file.file_variable_names = dealii::Utilities::split_string_list(
-              parameter_handler.get("file variable names"));
-            ic_file.simulation_variable_names = dealii::Utilities::split_string_list(
-              parameter_handler.get("simulation variable names"));
-            // Defaults to 0 for unused dimensions/cases that don't require it
-            for (unsigned int k = 0; k < dim; ++k)
-              {
-                ic_file.n_data_points.at(k) =
-                  static_cast<unsigned int>(parameter_handler.get_integer(
-                    "data points in " + axis_labels.at(k) + " direction"));
-              }
-            load_ic_parameters.add_initial_condition_file(ic_file);
-          }
-      }
-      parameter_handler.leave_subsection();
-    }
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_nucleation_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  parameter_handler.enter_subsection("nucleation");
-  {
-    nucleation_parameters.set_exclusion_distance(
-      parameter_handler.get_double("nucleus exclusion distance"));
-
-    nucleation_parameters.set_same_field_exclusion_distance(
-      parameter_handler.get_double("same field nucleus exclusion distance"));
-
-    nucleation_parameters.set_nucleation_period(
-      static_cast<unsigned int>(parameter_handler.get_integer("nucleation period")));
-
-    nucleation_parameters.set_refinement_radius(
-      parameter_handler.get_double("refinement radius"));
-
-    nucleation_parameters.set_seeding_time(parameter_handler.get_double("seeding time"));
-
-    nucleation_parameters.set_seeding_increments(
-      static_cast<unsigned int>(parameter_handler.get_integer("seeding increments")));
-  }
-  parameter_handler.leave_subsection();
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::assign_miscellaneous_parameters(
-  dealii::ParameterHandler &parameter_handler)
-{
-  parameter_handler.enter_subsection("miscellaneous");
-  {
-    misc_parameters.set_random_seed(static_cast<unsigned int>(
-      parameter_handler.get_integer("random seed") +
-      dealii::Utilities::MPI::this_mpi_process(MPI_COMM_WORLD)));
-  }
-  parameter_handler.leave_subsection();
-}
-
-template <unsigned int dim>
-void
-UserInputParameters<dim>::load_model_constants(
-  const InputFileReader    &input_file_reader,
-  dealii::ParameterHandler &parameter_handler)
-{
-  for (const std::string &constant_name : input_file_reader.get_names())
-    {
-      std::string constants_text = "Model constant ";
-      constants_text.append(constant_name);
-
-      std::vector<std::string> model_constants_strings =
-        dealii::Utilities::split_string_list(parameter_handler.get(constants_text));
-
-      user_constants.add_user_constant(constant_name, model_constants_strings);
-    }
-}
 
 #include "user_inputs/user_input_parameters.inst"
